requirements.txt
# Core Azure + OpenAI SDKs
openai~=1.60.2
azure-identity~=1.20.0
ansible-core~=2.17.0

# Data & analytics stack for fraud detection
pandas~=2.2.2
numpy~=1.26.4
scikit-learn~=1.5.2
python-dateutil~=2.9.0.post0

# Optional: text utils & parsing
tqdm~=4.66.5